Rename WebpackAssetPipeline -> ExternalAssetPipeline
We started using gulp in addition to webpack to compile these assets, so
calling the module WebpackAssetPipeline is no longer appropriate. What's
common to gulp and webpack is that they are independently managed tools
that provide a revisioned asset manifest, which we've integrated with
rails. We've effectively created an external asset pipeline, so the name
ExternalAssetPipeline seems appropriate.