Make ExternalAssetPipeline#path_to_asset respect options[:skip_pipeline]
From the rails documentation [1]:

> When using the asset pipeline (i.e. sprockets and sprockets-rails), the
> behavior is "enhanced". You can bypass the asset pipeline by passing in
> <tt>skip_pipeline: true</tt> to the options.

We stopped using sprockets after replacing its functionality with the
ExternalAssetPipeline and nodejs tooling. However, ExternalAssetPipeline
would _always_ look in its manifest, even when it was unnecessary (e.g.
when passing a full path to an asset in the public directory, or when
passing a base64-encoded data-URI string). To solve that, we're adding
support for the `skip_pipeline` option in the ExternalAssetPipeline.
Now, if you call an asset tag helper like `javascript_include_tag` with
the `skip_pipeline` option, it won't look in the manifest (it will just
pass it on directly to ActionView's base `path_to_asset` method).

[1]: https://github.com/rails/rails/blob/v5.2.2.1/actionview/lib/action_view/helpers/asset_url_helper.rb#L123-L125