Refactor ExternalAssetPipeline to override `compute_asset_path`
instead of `path_to_asset`

The way sprockets-rails hooks into rails asset tag helpers is by
overriding the `compute_asset_path` method:
https://github.com/rails/sprockets-rails/blob/v3.2.1/lib/sprockets/rails/helper.rb#L74-L96

This is what is referred to in the rails documentation [1] when it says:
> When using the asset pipeline (i.e. sprockets and sprockets-rails), the
> behavior is "enhanced"

It seems appropriate to make the ExternalAssetPipeline follow this
blueprint for hooking into/"enhancing" the asset tag helpers
(particularly because the existence of this established pattern makes
it pretty stable: changing the way this works in the rails repository
would require separate coordinated changes in the sprockets-rails
repository).

But beyond that, this allows us to simplify the ExternalAssetPipeline:
- We no longer need to define the `pack_name_with_extension` method
  (which we had lifted from Webpacker:
  https://github.com/rails/webpacker/blob/v3.5.5/lib/webpacker/helper.rb#L89-L91)
  because ActionView's built-in implementation of `path_to_asset`
  (aka `asset_path`) already handles adding the extension when
  necessary.
- We no longer need to implement our own support for the `skip_pipeline`
  option because that's already handled by ActionView's built-in
  `asset_path` implementation.

In addition, this also lets us take advantage of the logic in the base
`asset_path` implementation from ActionView, such as [2]:
```
return source if URI_REGEXP.match?(source)
```
so that we can return early when appropriate without looking in the
manifest (e.g. when the argument is a full URL, or a base64-encoded
data-URI string, or an absolute path to an asset in the public
directory)

[1]: https://github.com/rails/rails/blob/v5.2.2.1/actionview/lib/action_view/helpers/asset_url_helper.rb#L123-L124
[2]: https://github.com/rails/rails/blob/v5.2.2.1/actionview/lib/action_view/helpers/asset_url_helper.rb#L188