Add fall_back_to_sprockets configuration attribute
This setting will toggle whether the ExternalAssetPipeline should raise
when it can't find an asset or fall back to looking for the asset in the
sprockets pipeline (which itself will raise for missing assets unless
the `unknown_asset_fallback` setting is enabled).

The fall_back_to_sprockets setting defaults to false so that it'll raise
when used in an application that doesn't use sprockets.