[demo_app] Integrate external_asset_pipeline railtie
The railtie is required after `rails/all` which includes
`sprockets/railtie`. This ensures that the ExternalAssetPipeline takes
priority over sprockets (but falls back to sprockets when assets are not
found in the ExternalAssetPipeline, as indicated by enabling the
`config.external_asset_pipeline.fall_back_to_sprockets` setting).

Additionally, we disable manifest caching in the development environment
so that the server won't need to be restarted to pick up changes in the
manifest during development.