Rename `test` rake task to `test:unit`
We're going to add an integration test to test the ExternalAssetPipeline
railtie, which will be through a separate `test:integration` rake task.
So this is just the initial preparation for having two test rake tasks.

The `test` rake task now runs the `test:unit` rake task, but soon it'll
also run the `test:integration` rake task.