Add public_origin attribute to Configuration::DevServerSettings
This is an _optional_ configuration attribute which will be used to
generate the URLs in asset tags. If this isn't set, the asset origin
defaults to `"http://#{dev_server.host}:#{dev_server.port}"`, so
`public_origin` only needs to be set if that default doesn't work. In
essence, this setting allows different origins to be used when the Rails
backend connects to the dev server to get the manifest and when the
browser requests assets from the dev server. (Often, though, the same
origin can be used, which is why this setting is optional.)