Rename demo_app-no-sprockets -> demo_app-gulp
After adding the `demo_app-brunch` example, we now have two example apps
that don't use sprockets. Thus, the name `demo_app-no-sprockets` was a
bit confusing. Since the difference between `demo_app-no-sprockets` and
`demo_app` is that the former uses gulp instead of sprockets, we've
decided to rename it to `demo_app-gulp`.