Remove ServerManifest#data override
It was never necessary to override the parent Manifest class's #data
method. While it's unlikely for the additional
`if @config.cache_manifest?` condition to be `true` when a
ServerManifest is in use, we shouldn't outright prevent that use case.
And when it isn't `true`, the logic becomes equivalent to what we'd
defined in the overridden method.