Add prepend_assets_prefix_to_manifest_values attribute to Configuration
This attribute will be used to support manifests whose values already
include the assets prefix (and thus do not need it to be prepended).
This setting defaults to `true` which aligns with the default value of
the WebpackAssetsManifest plugin's `publicPath` option, which is `null`:
https://github.com/webdeveric/webpack-assets-manifest/blob/v3.1.1/readme.md#publicpath

However, it can be set to `false` to support manifests where the values
include the prefix, like a manifest generated by WebpackAssetsManifest
with the `publicPath` option set to `true` or to a string of the assets
prefix. (Right now, though, this attribute is unused - but we'll
refactor the code to start using it in subsequent commits.)