Update rubocop configuration to support the latest release (v0.85.1)
The Metrics/LineLength cop was moved to Layout/LineLength, and its
default value for the `Max` attribute was changed from 80 to 120. We're
going to stick to 80 for now, though (may revisit later).

Additionally, we've set `NewCops: enable` to opt into new cops. If this
becomes a pain we can instead switch to enabling individual cops, but
for now we'd like to stay as close to the default style guide as
possible.