Update instructions in README to be compatible with rails 7
It's important that the ExternalAssetPipeline railtie is loaded *after*
the Sprockets railtie in an application that uses Sprockets. Otherwise,
the `compute_asset_path` method from sprockets-rails will be invoked
prior to the one from `ExternalAssetPipeline::Helper`, which will lead
to an error when it attempts to resolve assets from the external asset
pipeline.

In Rails 7.0 (and beyond) sprockets-rails is no longer an explicit
dependency, so its railtie will be loaded via
`Bundler.require(*Rails.groups)` (rather than via `require "rails/all"`)
when sprockets-rails (or a gem which depends on sprockets-rails) is
explicitly declared in the Gemfile. Thus, we should place our
`require "external_asset_pipeline/railtie"` after
`Bundler.require(*Rails.groups)` so that it'll take precedence over
sprockets-rails no matter which version of rails is used.